+++ NEWS.md
@@ -1,3 +1,9 @@
+# testthatmulti 0.2.0
+
+* Added ttm_expect_error() and ttm_expect_no_error()
+
+* Fixed when referring to outside variables
+
 # testthatmulti 0.1.0
 
-* Initial CRAN submission.
+* Initial CRAN submission. Accepted to CRAN on 2/5/2024.